src/mcp-invoice-data/pyproject.toml
[project]
name = "mcp-invoice-data"
version = "0.1.0"
description = "MCP Server for invoice and purchase order data"
requires-python = ">=3.11"
dependencies = [
    "azure-monitor-opentelemetry>=1.6.0",
    "fastmcp>=2.0.0",
    "opentelemetry-instrumentation-fastapi>=0.50b0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "uvicorn>=0.30.0",
    "azure-ai-projects==2.0.0b2",
    "azure-identity==1.25.1",
    "openai>=2.8.1",
    "azure-ai-agents>=1.1.0",
    "aiohttp>=3.13.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
